move lodash -> merge
