+++ superset-frontend/packages/superset-ui-core/src/utils/index.ts
@@ -33,4 +33,4 @@ export * from './random';
 export * from './typedMemo';
 export * from './html';
 export * from './tooltip';
-export * from './lodash';
+export * from './merge';
